requirements.txt
fastapi[standard]
uvicorn
alembic
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
python-dotenv